pyproject.toml
[tool.poetry]
name = "elevenlabs-hackathon-25"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
elevenlabs = {extras = ["pyaudio"], version = "^1.51.0"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
